requirements.txt
opencv-contrib-python
ultralytics
paddleocr
paddlepaddle
thefuzz
onnxruntime
scipy
numpy
Pillow